Implement half-price child discounts for all MAG events (#10)
* Implement half-priced child badges for all MAG events
This code was in the magprime plugin, but both Stock and West leadership teams have requested this functionality so I'm copying it into our generic MAGFest plugin.

* Add child-related validations
Moves validations over from the magprime repo. We do not include the validations regarding c.CHILD_BADGE because we do not use the Child badge type in other events (yet?).

* Use age discount instead of cost property
The age discount is figured into the `badge_cost` property, while the cost_property decorators are applied afterwards. Since we're dealing with the badge cost, this is the better way to do it -- for example, the admin page will now show the correct badge price for an attendee under 13.

* Add child price tests
Also allows a configured age discount to 'win' over half-off prices, so an event can make child badges cheaper than the standard half-off discount without code changes.

* Fix unit tests

* Stop incurring false positive on one unit tests